backend/python-grpc/requirements.txt
grpcio
grpcio-tools
SQLAlchemy
psycopg2-binary
prometheus_client
opentelemetry-api
opentelemetry-sdk
sentry-sdk
sphinx
hupper
alembic